few tweaks

+++ src/main/java/dk/easv/easvexam/dal/ProfilesDAO.java
@@ -2,6 +2,7 @@ package dk.easv.easvexam.dal;
 
 import dk.easv.easvexam.be.MyException;
 import dk.easv.easvexam.be.Profile;
+import dk.easv.easvexam.gui.OpenView;
 
 import java.sql.*;
 import java.util.ArrayList;
@@ -61,6 +62,9 @@ public class ProfilesDAO {
     }
 
     public void deleteProfile(int id) throws MyException {
+        if (id == 11) {
+            throw new MyException("System error: The Standard Profile is system-critical and cannot be deleted!", null);
+        }
         try (Connection con = cm.getConnection()) {
             String sql = "DELETE FROM Profiles WHERE id = ?";
             PreparedStatement pstmt = con.prepareStatement(sql);
